+++ app/lib/l10n/app_localizations_da.dart
@@ -1205,6 +1205,9 @@ class AppLocalizationsDa extends AppLocalizations {
   String get commonFormationGrid => 'Grid';
 
   @override
+  String get commonFormationQuadruplet => 'Quadruplet';
+
+  @override
   String get commonFormationOther => 'Anden';
 
   @override