+++ app/lib/l10n/app_localizations_en.dart
@@ -1197,6 +1197,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get commonFormationGrid => 'Grid';
 
   @override
+  String get commonFormationQuadruplet => 'Quadruplet';
+
+  @override
   String get commonFormationOther => 'Other';
 
   @override