+++ packages/compendium_core/lib/src/imports/callersbox_adapter.dart
@@ -1207,6 +1207,8 @@ class CallersBoxAdapter implements SourceAdapter {
       shape = FormationShape.grid;
     } else if (lower.contains('circle')) {
       shape = FormationShape.circleMixer;
+    } else if (lower.contains('quadruplet')) {
+      shape = FormationShape.quadruplet;
     } else if (lower.contains('longways')) {
       shape = FormationShape.longways;
     }