+++ packages/compendium_core/lib/src/imports/contradb_html_adapter.dart
@@ -412,6 +412,8 @@ class ContraDbHtmlAdapter implements SourceAdapter {
       shape = FormationShape.grid;
     } else if (lower.contains('circle')) {
       shape = FormationShape.circleMixer;
+    } else if (lower.contains('quadruplet')) {
+      shape = FormationShape.quadruplet;
     } else if (lower.contains('longways')) {
       shape = FormationShape.longways;
     }